Initial lambda function
This files holds the code that is placed on aws.amazon.com as a Lambda function and linked to the Alexa skill. The full code is the .zip you download in the tutorial. The code just randomly selects a response for each intent.
The ARN should be placed in APP_ID.